refactor: Add new demo files
