--- examples/demo/main.py
@@ -1,4 +1,0 @@
-import demo
-
-if __name__ == "__main__":
-    demo.main()
