--- examples/minimal/main.py
@@ -1,4 +1,0 @@
-import minimal
-
-if __name__ == "__main__":
-    minimal.main()
