--- examples/plot/main.py
@@ -1,4 +1,0 @@
-import plot
-
-if __name__ == "__main__":
-    plot.main()
